Close #6: ,r to run the current file
`,r` runs the current file in a specified tmux pane. It calls a function
I defined, which stores the value of the target pane after the first
time it asks for it.

If you want to change the tmux target pane, run
`:ChangeTmuxPaneForFileExecution`.

The function also checks for the `g:execute_with` variable: it it's not
set (inside some autocmd) then it prompts for it too.