Fix a bug in the 'l' alias on Linux
Close #18

The `l` alias (`ls` with some options) used the `-G` option along with other
options and a dynamic option for coloring the output, which was set to `-G` on
OSX and `--color` on GNU ls. Adding the `-G` options non-dynamically caused the
dynamic option to be ignored and caused linux (Ubuntu server in this case) to
omit the group in the file listing.

I just removed the `-G` static flag from the alias and everything works fine.